feat:created two modals, Style: update design and other fixes
